Almost all remaing occurences of "links" changed to "elinks". Produced binary is named "elinks" now, global config directory is /etc/elinks/ now, local config directory is ~/.elinks/ now, links.cfg is elinks.cfg now, links.his is gotohist now, history is globhist now. Plenty of minor changes and updates accompanied.
YOU DO NOT WANT TO USE BUILD FROM NOW!!! Current configuration is lost for new
ELinks, as it stays in ~/.links and ELinks will look for it in ~/.elinks now.
But you do NOT want to clone it! Format of some config files (at least
bookmarks and .cfg) is going to change in next days, so you will just confuse
your new ELinks with old configuration files there. Whatever you will put in
~/.elinks may disappear randomly in next days. I'll tell you when we'll be
over ;).

*Yawn* and now I'm going to watch GitS ;).