The first two fixes keep old form data around and avoid crashing on null value fields.
The last fix is my crude attempt to delete forms in and outside of lists
without underflow/overflow. There should probably be a free_list_with_callback
version that calls a per-element freer on each list member to reduce the number
of times that lists are freed incorrectly.

Patch by Zardoz.