Implement support for scrollable text widgets. First usage is the Keys message box. It is not perfect yet but has not segfaulted for a long time now and I would like to have a reference for further hacking.
Text widgets can now be focusable but only when there is actually anything to scroll.
This is checked after calling the layouter whether we should cycle focus.

Forcing redrawing while the textwidget is scrollable will resize the dialog (and lead
to unwanted recalculations of the line splitting. This should maybe be fixed. Also
I'd like to have some percentage overview (like the title bar) when we are scrolling.