Instead of zeroing the session's 'visited' flag in display_tab_bar, zero it in ses_forward.
This is more logical and efficient, as we now zero the flag once,
when we finish loading the page, instead of repeatedly,
once every time we update the tab bar, from when we first start trying
to load the page to when we finish or fail.

More importantly, it fixes a bug whereby we would leave
the tab highlighted as freshly loaded even if the page failed to load.