Support for cursor routing. It should be fairly complete already altho' we should ofcourse get some feedback from an expert. It is based on patch by Witold Filipczyk, available at [1], which introduces cursor moving actions.
The session struct maintains whether we are doing cursor routing so setting the
terminal cursor and window pointer is done correctly. Also, current link is
only updated by check_vs() if navigating link-wise.

When in cursor routing mode, the cursor position is maintained in the session
tab and restored from it after things like doing the link menu. All cursor
updating is done in new move_cursor() function which updates current link when
cursor is over one and scrolls when moving outside the view.

[1] http://archive.linuxfromscratch.org/mail-archives/elinks-dev/2003-November/000596.html