get_extension_from_uri: Include the leading '.' in the returned extension.
get_temp_name: Now that get_extension_from_uri includes the '.',
do not add another here.

get_content_type_mimetypes: Increment |extension| before the first iteration
to skip the leading '.'. The loop iteratively truncates the beginning
of the string up to and including the first '.' and looks up the result,
and so the routine would still work without this change, but it looks better
with the increment and avoid a first interation in which we would look up
the content type for an empty extension.

The impetus for this change was to fix the code in http_got_header that detects
whether the content encoding matches the extension of the file, which is
the heuristic used to determine whether to decompress the data.  http_got_header
passes the result of |get_extension_from_uri(uri)| to guess_encoding,
which expects the leading '.'. Thus, we were always decoding the data, and now
we again decode only when the extension does not match the content encoding
(e.g., 'Content-Encoding: gzip' and file name 'foo.tar.gz').