Add a new parameter |convert_slashes| to encode_uri_string to indicate whether slashes should be converted along with other unsafe characters to hexadecimal notation.
When prepending "file://" to the string to convert a path name to a URI
in encode_uri_string, also encode the path name.

Now, when a user provides on the command line or in the Go to URL box
a path with question marks, octothorpes, or other characters considered
unsafe, they will be encoded so that later on we do not misinterpret
part of the file name as a fragment identifier, query part, or whatever.

This fixes bug 358: browsing local files fails if path contains "%7E".