In add_document_line, check whether |expanded + line_pos| is non-zero rather than maintaining a count of backspaces and checking whether it doubled is less than |line_pos|.
This will affect the current quirky behaviour on tabs followed by backspaces:
Now, we will avoid backspacing beyond the start of the expanded line rather
than beyond the start of the line as counted in characters.  I intend
eventually to improve backspace handling so that a backspace that immediately
follows a tab will backspace the whole tab in the first place.