It's bigger, yes, but, I believe, much cleaner.
You can now enter "search terms" (with the quotes) in the goto box and
it will use your default search engine (or Google).  Use ', `, or ".

New config.pl option: "language".  Set this and you no longer have to
enter the "to language" (although you still can) when translating
a page.

Added eff, wired, and slashdot as "news" sources.  Set the config option
"news: eff" and the prefixes "news" and "n" will use this.

There are more config options, and they are now documented towards the
bottom of the script in a commented sample config.pl.

                                                                    Russ